requirements.txt
alabaster==0.7.11
alembic==1.0.7
altair==4.0.0
anaconda-client==1.7.2
anaconda-navigator==1.6.8
anaconda-project==0.8.2
appconfigpy==0.0.7
appdirs==1.4.3
apyori==1.1.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backtrader==1.9.70.122
base58==1.0.3
bcolz==0.12.1
bcrypt==3.1.6
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
blinker==1.4
bokeh==0.13.0
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
Bottleneck==1.2.1
BTrees==4.5.1
CacheControl==0.12.3
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.8.0
conda-build==3.0.22
conda-package-handling==1.6.0
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.5
CurrencyConverter==0.13.11
cycler==0.10.0
cymem==2.0.2
cyordereddict==1.0.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
DataProperty==0.36.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.9
distlib==0.2.5
distributed==1.23.1
docopt==0.6.2
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
empyrical==0.5.0
en-core-web-sm==2.0.0
entrypoints==0.2.3
enum-compat==0.0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.0.2
feedparser==5.2.1
filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
forex-python==1.5
future==0.18.2
gevent==1.3.6
glob2==0.6
greenlet==0.4.15
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
IbPy2==0.8.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==1.3.0
incremental==17.5.0
inflection==0.3.1
intervaltree==3.0.2
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.4
json5==0.8.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.0
kaggle==1.5.0
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lockfile==0.12.2
Logbook==1.4.1
lru-dict==1.1.6
lxml==4.2.5
Mako==1.0.7
markdown2==2.3.6
MarkupSafe==1.0
matplotlib==2.2.3
mbstrdecoder==0.5.0
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
mlxtend==0.15.0.0
more-itertools==4.3.0
mpmath==1.0.0
msgfy==0.0.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multipledispatch==0.6.0
murmurhash==1.0.2
navigator-updater==0.1.0
nbconvert==5.4.0
nbformat==4.4.0
ndg-httpsclient==0.4.3
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.16.1
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
packaging==17.1
pandas==0.23.4
pandas-datareader==0.6.0
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
pathtools==0.1.2
pathvalidate==0.21.4
patsy==0.5.0
pep8==1.7.1
persistent==4.4.3
pickleshare==0.7.4
Pillow==7.0.0
pipreqs==0.4.10
pkginfo==1.4.2
plac==0.9.6
plotly==4.4.1
pluggy==0.7.1
ply==3.11
preshed==2.0.1
progress==1.3
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.11.2
psutil==5.4.7
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
PyCurrency-Converter==1.0
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt5-sip==4.19.14
pyreadline==2.1
pyrsistent==0.15.6
PySocks==1.6.8
pytablereader==0.22.2
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-editor==1.0.4
python-slugify==1.2.6
pytz==2018.5
PyWavelets==1.0.0
pywin32==223
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
Quandl==3.3.0
regex==2018.1.10
requests==2.20.1
requests-file==1.4.3
requests-ftp==0.3.1
retrying==1.3.3
rope==0.11.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.16.0
SimpleSQLite==0.33.2
singledispatch==3.4.0.3
sip==4.19.6
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
soupsieve==1.9.5
spacy==2.0.18
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
sqlitebiter==0.23.2
sqliteschema==0.13.4
statsmodels==0.9.0
streamlit==0.52.2
sympy==1.2
tabledata==0.0.25
tables==3.4.4
TBB==0.1
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
thinc==6.12.1
toml==0.10.0
toolz==0.9.0
tornado==5.1
tqdm==4.28.1
trading-calendars==1.7.0
traitlets==4.3.2
transaction==2.4.0
Twisted==18.7.0
typed-ast==1.1.0
typepy==0.3.0
typing==3.6.2
tzlocal==2.0.0
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.23
validators==0.14.1
vega-datasets==0.8.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
wxPython==4.0.4
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
yahoo-finance==1.4.0
yarg==0.1.9
zc.lockfile==1.4
ZConfig==3.4.0
zict==0.1.3
zipline==1.3.0
zipp==0.6.0
ZODB==5.5.1
zodbpickle==1.0.3
zope.interface==4.6.0
